+++ private/buf/cmd/buf/command/beta/registry/plugin/plugincreate/plugincreate.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/bufbuild/buf/private/buf/bufcli"
 	"github.com/bufbuild/buf/private/buf/bufprint"
-	"github.com/bufbuild/buf/private/bufpkg/bufplugin"
+	"github.com/bufbuild/buf/private/bufpkg/bufremoteplugin"
 	registryv1alpha1 "github.com/bufbuild/buf/private/gen/proto/go/buf/alpha/registry/v1alpha1"
 	"github.com/bufbuild/buf/private/pkg/app/appcmd"
 	"github.com/bufbuild/buf/private/pkg/app/appflag"
@@ -49,7 +49,7 @@ func NewCommand(
 ) *appcmd.Command {
 	flags := newFlags()
 	return &appcmd.Command{
-		Use:   name + " <buf.build/owner/" + bufplugin.PluginsPathName + "/plugin>",
+		Use:   name + " <buf.build/owner/" + bufremoteplugin.PluginsPathName + "/plugin>",
 		Short: "Create a new Protobuf plugin.",
 		Args:  cobra.ExactArgs(1),
 		Run: builder.NewRunFunc(
@@ -106,7 +106,7 @@ func run(
 	if err != nil {
 		return err
 	}
-	remote, owner, name, err := bufplugin.ParsePluginPath(pluginPath)
+	remote, owner, name, err := bufremoteplugin.ParsePluginPath(pluginPath)
 	if err != nil {
 		return err
 	}