+++ private/buf/cmd/buf/command/push/push.go
@@ -21,7 +21,6 @@ import (
 	"github.com/bufbuild/buf/private/buf/bufcli"
 	"github.com/bufbuild/buf/private/bufpkg/bufanalysis"
 	"github.com/bufbuild/buf/private/bufpkg/bufmodule"
-	"github.com/bufbuild/buf/private/bufpkg/bufmodule/bufmoduleref"
 	"github.com/bufbuild/buf/private/pkg/app/appcmd"
 	"github.com/bufbuild/buf/private/pkg/app/appflag"
 	"github.com/bufbuild/buf/private/pkg/command"
@@ -32,11 +31,12 @@ import (
 )
 
 const (
-	trackFlagName           = "track"
 	tagFlagName             = "tag"
 	tagFlagShortName        = "t"
 	errorFormatFlagName     = "error-format"
 	disableSymlinksFlagName = "disable-symlinks"
+	// deprecated
+	trackFlagName = "track"
 )
 
 // NewCommand returns a new Command.
@@ -61,10 +61,11 @@ func NewCommand(
 }
 
 type flags struct {
-	Tracks          []string
 	Tags            []string
 	ErrorFormat     string
 	DisableSymlinks bool
+	// Deprecated
+	Tracks []string
 	// special
 	InputHashtag string
 }
@@ -76,12 +77,6 @@ func newFlags() *flags {
 func (f *flags) Bind(flagSet *pflag.FlagSet) {
 	bufcli.BindInputHashtag(flagSet, &f.InputHashtag)
 	bufcli.BindDisableSymlinks(flagSet, &f.DisableSymlinks, disableSymlinksFlagName)
-	flagSet.StringSliceVar(
-		&f.Tracks,
-		trackFlagName,
-		nil,
-		"Append the pushed module to this track. Multiple tracks are appended if specified multiple times.",
-	)
 	flagSet.StringSliceVarP(
 		&f.Tags,
 		tagFlagName,
@@ -98,6 +93,13 @@ func (f *flags) Bind(flagSet *pflag.FlagSet) {
 			stringutil.SliceToString(bufanalysis.AllFormatStrings),
 		),
 	)
+	flagSet.StringSliceVar(
+		&f.Tracks,
+		trackFlagName,
+		nil,
+		"Do not use. This flag never had any effect",
+	)
+	_ = flagSet.MarkHidden(trackFlagName)
 }
 
 func run(
@@ -105,6 +107,9 @@ func run(
 	container appflag.Container,
 	flags *flags,
 ) (retErr error) {
+	if len(flags.Tracks) > 0 {
+		return appcmd.NewInvalidArgumentErrorf("--%s has never had any effect, do not use.", trackFlagName)
+	}
 	if err := bufcli.ValidateErrorFormatFlag(flags.ErrorFormat, errorFormatFlagName); err != nil {
 		return err
 	}
@@ -138,10 +143,6 @@ func run(
 	if err != nil {
 		return err
 	}
-	tracks := flags.Tracks
-	if tracks == nil {
-		tracks = []string{bufmoduleref.MainTrack}
-	}
 	localModulePin, err := service.Push(
 		ctx,
 		moduleIdentity.Owner(),
@@ -149,7 +150,7 @@ func run(
 		"",
 		protoModule,
 		flags.Tags,
-		tracks,
+		nil,
 	)
 	if err != nil {
 		if rpc.GetErrorCode(err) == rpc.ErrorCodeAlreadyExists {